requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.11.0
django-filter==21.1
django-phone-field==1.8.1
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.29
PyJWT==2.3.0
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.8.0
